+++ .github/workflows/go.yaml
@@ -55,7 +55,7 @@ jobs:
     environment: golang
     outputs:
       steampipe: ${{ steps.build_services.outputs.steampipe }}
-      auth-v2-service: ${{ steps.build_services.outputs.auth-v2-service }}
+      auth-service: ${{ steps.build_services.outputs.auth-service }}
       checkup-worker: ${{ steps.build_services.outputs.checkup-worker }}
       compliance-report-worker: ${{ steps.build_services.outputs.compliance-report-worker }}
       compliance-service: ${{ steps.build_services.outputs.compliance-service }}
@@ -181,7 +181,7 @@ jobs:
           build-args: |
             PLUGIN_REGISTRY=ghcr.io/opengovern
           context: .
-  deploy-auth-v2-service:
+  deploy-auth-service:
     runs-on: ubuntu-latest
     needs:
       - build
@@ -190,7 +190,7 @@ jobs:
       id-token: write
       contents: read
     environment: docker
-    if: needs.build.outputs.auth-v2-service == 'true' && github.event_name != 'pull_request' && (github.ref == 'refs/heads/main' || github.ref == 'refs/heads/dev')
+    if: needs.build.outputs.auth-service == 'true' && github.event_name != 'pull_request' && (github.ref == 'refs/heads/main' || github.ref == 'refs/heads/dev')
     steps:
       - name: Checkout code
         uses: actions/checkout@v3
@@ -213,8 +213,8 @@ jobs:
         with:
           push: true
           tags: |
-            ghcr.io/${{ github.repository_owner }}/auth-v2-service:${{ needs.tag.outputs.latest_tag }}
-          file: docker/AuthV2ServiceDockerfile
+            ghcr.io/${{ github.repository_owner }}/auth-service:${{ needs.tag.outputs.latest_tag }}
+          file: docker/AuthServiceDockerfile
           context: .
   deploy-checkup-worker:
     runs-on: ubuntu-latest
