+++ build_services
@@ -1,4 +1,4 @@
-authV2-service
+auth-service
 checkup-worker
 cli
 cloud-inventory
