+++ cmd/auth-service/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"os/signal"
 	"syscall"
-	"github.com/opengovern/opengovernance/pkg/authV2"
+	"github.com/opengovern/opengovernance/pkg/auth"
 )
 
 func main() {
@@ -28,7 +28,7 @@ func main() {
 		}
 	}()
 
-	if err := authV2.Command().ExecuteContext(ctx); err != nil {
+	if err := auth.Command().ExecuteContext(ctx); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
