+++ docker/AuthServiceDockerfile
@@ -3,5 +3,5 @@ RUN apk --no-cache add ca-certificates
 
 FROM scratch
 COPY --from=build /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
-COPY ./build/auth-v2-service /
-CMD [ "/auth-v2-service" ]
+COPY ./build/auth-service /
+CMD [ "/auth-service" ]
