+++ pkg/auth/cmd.go
@@ -1,4 +1,4 @@
-package authV2
+package auth
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 	config2 "github.com/opengovern/og-util/pkg/config"
 	"github.com/opengovern/og-util/pkg/httpserver"
 	"github.com/opengovern/og-util/pkg/postgres"
-	"github.com/opengovern/opengovernance/pkg/authV2/db"
+	"github.com/opengovern/opengovernance/pkg/auth/db"
 
 	"crypto/rand"
 
@@ -24,7 +24,7 @@ import (
 )
 
 var (
-	dexAuthDomain         = os.Getenv("DEX_AUTH_DOMAIN")
+dexAuthDomain         = os.Getenv("DEX_AUTH_DOMAIN")
 	dexAuthPublicClientID = os.Getenv("DEX_AUTH_PUBLIC_CLIENT_ID")
 	dexGrpcAddress        = os.Getenv("DEX_GRPC_ADDR")
 
@@ -62,7 +62,7 @@ func start(ctx context.Context) error {
 		return err
 	}
 
-	logger = logger.Named("authV2")
+	logger = logger.Named("auth")
 
 	dexVerifier, err := newDexOidcVerifier(ctx, dexAuthDomain, dexAuthPublicClientID)
 	if err != nil {
@@ -74,7 +74,7 @@ func start(ctx context.Context) error {
 
 	// setup postgres connection
 	cfg := postgres.Config{
-		Host:    conf.PostgreSQL.Host,
+Host:    conf.PostgreSQL.Host,
 		Port:    conf.PostgreSQL.Port,
 		User:    conf.PostgreSQL.Username,
 		Passwd:  conf.PostgreSQL.Password,
