+++ pkg/metadata/api/metadata.go
@@ -1,10 +1,11 @@
 package api
 
 import (
-	authApi "github.com/opengovern/opengovernance/pkg/authV2/api"
+	"time"
+
+	authApi "github.com/opengovern/opengovernance/pkg/auth/api"
 	"github.com/opengovern/opengovernance/pkg/onboard/api"
 	"github.com/opengovern/opengovernance/services/migrator/db/model"
-	"time"
 )
 
 type SetConfigMetadataRequest struct {
