+++ services/migrator/job/migration_list.go
@@ -2,7 +2,6 @@ package job
 
 import (
 	"github.com/opengovern/opengovernance/services/migrator/job/migrations/analytics"
-	"github.com/opengovern/opengovernance/services/migrator/job/migrations/authV2"
 	"github.com/opengovern/opengovernance/services/migrator/job/migrations/compliance"
 	"github.com/opengovern/opengovernance/services/migrator/job/migrations/elasticsearch"
 	"github.com/opengovern/opengovernance/services/migrator/job/migrations/inventory"
@@ -11,6 +10,8 @@ import (
 	"github.com/opengovern/opengovernance/services/migrator/job/migrations/resource_info"
 	"github.com/opengovern/opengovernance/services/migrator/job/migrations/workspace"
 	"github.com/opengovern/opengovernance/services/migrator/job/types"
+		"github.com/opengovern/opengovernance/services/migrator/job/migrations/auth"
+
 )
 
 var migrations = map[string]types.Migration{
@@ -22,5 +23,5 @@ var migrations = map[string]types.Migration{
 	"compliance":          compliance.Migration{},
 	"analytics":           analytics.Migration{},
 	"resource_info":       resource_info.Migration{},
-	"authV2":              authV2.Migration{},
+	"auth":              auth.Migration{},
 }
